feat. detail 완료
