--- src/pages/Detail.jsx
@@ -1,77 +1,0 @@
-import React from 'react';
-import Header from '../common/Header';
-import Container from '../common/Container';
-import { useNavigate, useParams } from 'react-router-dom';
-
-export default function Detail() {
-  const params = useParams();
-
-  const navigate = useNavigate();
-  return (
-    <>
-      <Header />
-      <Container>
-        <h1
-          style={{
-            border: '1px solid lightgray',
-            borderRadius: '12px',
-            padding: '12px'
-          }}
-        >
-          제목
-        </h1>
-        <div
-          style={{
-            height: '400px',
-            border: '1px solid lightgray',
-            borderRadius: '12px',
-            padding: '12px'
-          }}
-        >
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Ad doloribus blanditiis vitae sapiente. Expedita
-          delectus nihil animi pariatur, labore quod officiis dolor fugit. Mollitia quod, delectus velit deleniti nihil
-          veniam!
-        </div>
-        <div
-          style={{
-            marginTop: '12px',
-            display: 'flex',
-            justifyContent: 'end'
-          }}
-        >
-          <button
-            onClick={() => {
-              navigate('/edit');
-            }}
-            style={{
-              border: 'none',
-              padding: '8px',
-              borderRadius: '6px',
-              backgroundColor: 'orange',
-              color: 'white',
-              cursor: 'pointer',
-              marginRight: '6px'
-            }}
-          >
-            수정
-          </button>
-          <button
-            onClick={() => {
-              alert('삭제할까?');
-            }}
-            style={{
-              border: 'none',
-              padding: '8px',
-              borderRadius: '6px',
-              backgroundColor: 'red',
-              color: 'white',
-              cursor: 'pointer'
-            }}
-          >
-            삭제
-          </button>
-        </div>
-      </Container>
-    </>
-  );
-}
