+++ src/pages/detail/Detail.styled.js
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+export const TitleLayout = styled.h2`
+  border: 1px solid lightgray;
+  border-radius: 12px;
+  padding: 12px;
+  margin-bottom: 16px;
+`;
+
+export const ContentLayout = styled.div`
+  height: 400px;
+  border: 1px solid lightgray;
+  border-radius: 12px;
+  padding: 12px;
+`;
