+++ src/shared/Router.js
@@ -1,6 +1,6 @@
 import { Route, Routes } from 'react-router-dom';
 import Main from '../pages/main/Main';
-import Detail from '../pages/Detail';
+import Detail from '../pages/detail/Detail';
 import Create from '../pages/Create';
 import Edit from '../pages/Edit';
 import Login from '../pages/Login';
